pyproject.toml
[project]
name = "flump"
version = "0.1.7"
description = "A small utility program to flip the luminance of images and save the result to your clipboard (for MacOS and Windows)"
authors = [
    { name = "Seth Hinz", email = "sethhinz@me.com" }
]
dependencies = [
    "pyqt6>=6.7.1",
    "pillow>=10.4.0",
    "numpy>=2.0.1",
    "pyobjc>=10.3.1; platform_system == 'Darwin'",
    "pyobjc-core>=10.3.1; platform_system == 'Darwin'",
    "pyinstaller>=6.9.0",
    "pywin32>=306; platform_system == 'Windows'",
    "chromakey>=0.1.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
flump = "flump.flump:main"
build = "flump.build:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/flump"]
